% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.eval_group}
\alias{.eval_group}
\title{Evaluate .group constructs}
\usage{
.eval_group(input = NULL, groups = NULL, positions = NULL)
}
\arguments{
\item{input}{[\code{character(1)}]\cr table to reorganise.}

\item{groups}{[\code{list(3)}]\cr the groups-slot from a schema.}

\item{positions}{[\code{integerish(.)}]\cr the cell column or row that should
be adapted to groupings.}
}
\value{
the position of the evaluated position
}
\description{
Evaluate .group constructs
}
