% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.eval_sum}
\alias{.eval_sum}
\title{Evaluate .sum constructs}
\usage{
.eval_sum(input = NULL, groups = NULL, data = NULL)
}
\arguments{
\item{input}{[\code{character(1)}]\cr table to reorganise.}

\item{groups}{[\code{list(3)}]\cr the groups-slot from a schema.}

\item{data}{[\code{integerish(.)}]\cr the cell column or row that should be
adapted to groupings.}
}
\value{
the position of the evaluated position
}
\description{
Evaluate .sum constructs
}
