% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{.find}
\alias{.find}
\title{Determine row or column on the fly}
\usage{
.find(
  fun = NULL,
  pattern = NULL,
  col = NULL,
  row = NULL,
  invert = FALSE,
  relative = FALSE
)
}
\arguments{
\item{fun}{[\code{character(1)}]\cr function to identify columns or rows in
the input table on the fly.}

\item{pattern}{[\code{character(1)}]\cr character string containing a regular
expression to identify columns or rows in the input table on the fly.}

\item{col}{[\code{integerish(1)}]\cr optionally, in case this function should
only be applied to certain columns, provides this here.}

\item{row}{[\code{integerish(1)}]\cr optionally, in case this function should
only be applied to certain rows, provides this here.}

\item{invert}{[\code{logical(1)}]\cr whether or not the identified columns or
rows should be inverted, i.e., all other columns or rows should be
selected.}

\item{relative}{[\code{logical(1)}]\cr whether or not the values provided in
\code{col} or \code{row} are relative to the cluster position(s) or whether
they are absolute positions, i.e, refer to the overall table.}
}
\value{
the index values where the target was found.
}
\description{
Find the location of a variable not based on it's columns/rows, but based on
a regular expression or function
}
\details{
This functions is basically a wild-card for when columns or rows are
  not known ad-hoc, but have to be assigned on the fly. This can be very
  helpful when several tables contain the same variables, but the arrangement
  may be slightly different.
}
\section{How does this work}{
 The first step in using any schema is validating
  it via the function \code{\link{validateSchema}}. This happens by default
  in \code{\link{reorganise}}, but can also be done manually, for example
  when debugging complicated schema descriptions.

  In case that function encounters a schema that wants to find columns or
  rows on the fly via \code{.find}, it combines all cells of columns and all
  cells of rows into one character string and matches the regular expression
  or function on those. Columns/rows that have a match are returned as the
  respective column/row value.
}

\examples{
# use regular expressions to find cell positions
(input <- tabs2shift$clusters_messy)

schema <- setCluster(id = "territories",
                     left = .find(pattern = "comm*"), top = .find(pattern = "unit")) \%>\%
  setIDVar(name = "territories", columns = c(1, 1, 4), rows = c(2, 9, 9)) \%>\%
  setIDVar(name = "year", columns = 4, rows = c(3:6), distinct = TRUE) \%>\%
  setIDVar(name = "commodities", columns = c(1, 1, 4)) \%>\%
  setObsVar(name = "harvested", columns = c(2, 2, 5)) \%>\%
  setObsVar(name = "production", columns = c(3, 3, 6))

schema
validateSchema(schema = schema, input = input)

# use a function to find rows
(input <- tabs2shift$messy_rows)

schema <-
  setFilter(rows = .find(fun = is.numeric, col = 1, invert = TRUE)) \%>\%
  setIDVar(name = "territories", columns = 1) \%>\%
  setIDVar(name = "year", columns = 2) \%>\%
  setIDVar(name = "commodities", columns = 3) \%>\%
  setObsVar(name = "harvested", columns = 5) \%>\%
  setObsVar(name = "production", columns = 6)

reorganise(schema = schema, input = input)
}
