% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.getColTypes}
\alias{.getColTypes}
\title{Get the column types of a tibble}
\usage{
.getColTypes(input = NULL, collapse = TRUE)
}
\arguments{
\item{input}{[\code{data.frame(1)}]\cr table of which to get the column
types.}

\item{collapse}{[\code{logical(1)}]\cr whether or not to paste all column
types into one string.}
}
\description{
Get the column types of a tibble
}
