% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.groupTable}
\alias{.groupTable}
\title{Extract summarised data}
\usage{
.groupTable(schema = NULL, input = NULL)
}
\arguments{
\item{schema}{[\code{character(1)}]\cr the (validated) schema description of
\code{input}.}

\item{input}{[\code{character(1)}]\cr table to reorganise.}
}
\value{
a table where columns and rows are summarised
}
\description{
This function extracts data from a table that are summarised by applying a
schema description to it.
}
