% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.reportProblems}
\alias{.reportProblems}
\title{Catch and report problems in a schema description}
\usage{
.reportProblems(schema = NULL)
}
\arguments{
\item{schema}{[\code{character(1)}]\cr the schema description to check.}
}
\description{
This function checks the current setup of a schema and reports problems that
will lead to an error of \code{reorganise} if not fixed.
}
