% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectData.R
\name{.selectData}
\alias{.selectData}
\title{Select data chunks from a spreadsheet}
\usage{
.selectData(input = NULL, schema = NULL)
}
\arguments{
\item{input}{the raw table from which to select clusters}

\item{schema}{the schema description that is the basis to derive data.}
}
\value{
A list of the rows and columns in the spreadsheet that contain data
  plus the header and the data, per cluster.
}
\description{
This function builds a list of those chunks that contain the data.
}
