% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.spliceHeader}
\alias{.spliceHeader}
\title{Splice the header into the table}
\usage{
.spliceHeader(input, rows = NULL)
}
\arguments{
\item{input}{[\code{data.frame(1)}]\cr table of which the header should be
shifted into the table.}

\item{rows}{[\{integeris(1)]\cr the number of rows to shift into the table.}
}
\description{
Splice the header into the table
}
