% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.tidyVars}
\alias{.tidyVars}
\title{Match variables}
\usage{
.tidyVars(ids = NULL, obs = NULL)
}
\arguments{
\item{ids}{list of id variables}

\item{obs}{list of observed variables}
}
\value{
a symmetric list of variables (all with the same dimensions)
}
\description{
This function matches id and observed variables and reshapes them accordingly
}
