% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.updateFormat}
\alias{.updateFormat}
\title{Update the formating of a table}
\usage{
.updateFormat(input = NULL, schema = NULL)
}
\arguments{
\item{input}{[\code{character(1)}]\cr table to reorganise.}

\item{schema}{[\code{character(1)}]\cr the schema description of
\code{input}.}
}
\description{
This function updates the format of a table by applying a schema description
to it.
}
