% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_valid.R
\name{expect_valid_schema}
\alias{expect_valid_schema}
\title{Test for a valid schema}
\usage{
expect_valid_schema(x = NULL)
}
\arguments{
\item{x}{a schema to test.}
}
\description{
Test for a valid schema
}
