% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRowCol.R
\name{find_}
\alias{find_}
\alias{find_row}
\alias{find_col}
\alias{.find}
\title{Determine row or column on the fly}
\usage{
find_row(pattern, input = NULL)

find_col(pattern, input = NULL)

.find(pattern)
}
\arguments{
\item{pattern}{[\code{character(1)}]\cr character string containing a regular
expression to be matched in the \code{input} rows or columns.}

\item{input}{[\code{data.frame}]\cr optinally the input data-frame to match
the \code{pattern} with.}
}
\value{
the index values where \code{pattern} was found.
}
\description{
Find the location of a variable not based on it's columns/rows, but based on
a regular expression/name
}
\details{
These functions are basically a wild-card for when columns or rows
  are not known ad-hoc, but have to be assigned on the fly (for example when
  tables of a series are largely the same, but some variable(s) sit in
  different columns). When used in a schema description, \code{input} would
  be omitted, because it is assigned during the run of
  \code{\link{reorganise}} based on the specific table.
}
