% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getGroupVar}
\alias{getGroupVar}
\title{Extract cluster group variable}
\usage{
getGroupVar(input = NULL, schema = NULL)
}
\arguments{
\item{input}{[\code{character(1)}]\cr table to reorganise.}

\item{schema}{[\code{character(1)}]\cr the schema description of
\code{input}.}
}
\description{
This function extracts the cluster grouping variable from a table by applying
a schema description to it.
}
