% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDVars.R
\name{getIDVars}
\alias{getIDVars}
\title{Extract identifying variables}
\usage{
getIDVars(input = NULL, schema = NULL)
}
\arguments{
\item{input}{[\code{character(1)}]\cr table to reorganise.}

\item{schema}{[\code{character(1)}]\cr the schema description of
\code{input}.}
}
\description{
This function extracts the identifying variables from a table by applying a
schema description to it.
}
