% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{Derive metadata from a schema description}
\usage{
getMetadata(data = NULL, schema = NULL)
}
\arguments{
\item{data}{a list containing the values of \code{selectData}.}

\item{schema}{the schema description that is the basis to derive metadata.}
}
\description{
Derive metadata from a schema description
}
