% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObsVars.R
\name{getObsVars}
\alias{getObsVars}
\title{Extract observed variables}
\usage{
getObsVars(input = NULL, schema = NULL)
}
\arguments{
\item{input}{[\code{character(1)}]\cr table to reorganise.}

\item{schema}{[\code{character(1)}]\cr the schema description of
\code{input}.}
}
\description{
This function extracts the observed variables from a table by applying a
schema description to it.
}
