% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_tables}
\alias{make_tables}
\title{Make a tables from a pdf file}
\usage{
make_tables(
  pdf = NULL,
  pages = NULL,
  rows = "canvas",
  cols = "canvas",
  use.centroid = TRUE,
  merge = "_-_"
)
}
\arguments{
\item{pdf}{[character(1)][character]\cr}

\item{pages}{[integerish(1)][integer]\cr the range of pages to process, by
default it's all the pages.}

\item{rows}{[character(1)][character]|[integerish(1)][integer]\cr the way to estimate rows, possible
values are "canvas", "word_height" (the most frequent height) or an integer
value of the number of rows.}

\item{cols}{[character(1)][character]|[integerish(1)][integer]\cr the way to estimate columns, possible
values are "canvas", "median" or an integer value of the number of columns.}

\item{use.centroid}{[logical(1)][logical]\cr whether or not to identify the
position of text elements by it's centroid, or the minimum x and y value.}

\item{merge}{[character(1)][character]\cr symbol by which to merge when several
text boxes fall into the same cell.}
}
\value{
a rectangular table
}
\description{
extract information from a PDF table and turn it into a table
}
\examples{
# syr_2000 <- make_tables(pdf = "2000_Rain and land use.pdf")
# syr_2000[[4]] <- make_tables(pdf = "2000_Rain and land use.pdf",
#                              pages = 4, use.centroid = TRUE)
}
