% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSchema.R
\name{makeSchema}
\alias{makeSchema}
\title{Check a schema description}
\usage{
makeSchema(schema = NULL)
}
\arguments{
\item{schema}{[\code{list(.)}]\cr the list of schema information. This can
contain the lists \code{clusters}, \code{format}, \code{header} and
\code{variables}.}
}
\value{
An object of class \code{\link{schema}}.
}
\description{
This function checks whether the schema description is formally correct.
}
