% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{plot_table}
\alias{plot_table}
\title{Make a tables from a pdf file}
\usage{
plot_table(tab, merge = "_-_")
}
\arguments{
\item{tab}{[data.frame(.)][data.frame]\cr}

\item{merge}{[character(1)][character]\cr symbol by which to merge when several
text boxes fall into the same cell.}
}
\value{
diagnostic plot of a table.
}
\description{
extract information from a PDF table and turn it into a table
}
\examples{
# please check the vignette for examples
}
