% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record}
\alias{record}
\title{Record table properties}
\usage{
record(input, schema = NULL, ...)
}
\arguments{
\item{input}{[\code{data.frame(1)}]\cr table for which to record properties.}

\item{schema}{[\code{symbol(1)}]\cr the schema description that has been
saved in the global environment.}

\item{...}{[\code{various}]\cr property specific arguments; see Details.}
}
\description{
This functions records properties of tables and stores them in a schema
description in preparation of reorganisation.
}
\details{
Read the vignette for the basics of schema descriptions, a default
  schema description is stored in \code{\link{meta_default}}.
}
