% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reg_default}
\alias{reg_default}
\title{The default algorithm to register messy tables}
\format{The object of class list contains the two lists \code{body},
  describing information about the table bodies and \code{variables},
  describing the variables within each table.}
\usage{
reg_default
}
\description{
The default algorithm to register messy tables
}
\keyword{datasets}
