% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register}
\alias{register}
\title{Register table properties}
\usage{
register(input, what = NULL, export = FALSE, ...)
}
\arguments{
\item{input}{[\code{data.frame(1)}]\cr table for which to register
properties.}

\item{what}{[\code{character(1)}]\cr properties that should be registered as
deviating from the default. Possible values are \code{schema},
\code{format}, \code{clusters}, \code{commodities}, \code{adminUnits},
\code{years}.}

\item{export}{[\code{logical(1)}]\cr whether or not to export the meta data
as \code{XML}.}

\item{...}{[\code{various}]\cr property specific arguments; see Details.}
}
\description{
This functions registers properties of tables in preparation of
rectangularisation.
}
\details{
Property-specific arguments are:\itemize{ \item \code{schema}:
  \itemize{\item \code{name = ...}} A schema is a list of table
  properties. When calling it, \code{register} sets all of the properties
  specified in the schema. \item \code{format:} \itemize{ \item \code{type
  = "long" | "wide"} - if each variable is in one column, the table is long;
  if some variables are spread over several columns, the table is wide.}
  \item \code{clusters:} \itemize{\item \code{row} - the row number(s) of the
  top-left corner of each cluster. \item \code{col} - the column number(s) of
  the top-left corner of each cluster. \item \code{width} - the width of
  clusters (in number of cells) can either have one value (when all clusters
  have the same width) or as many values as there are clusters. \item
  \code{height} - the height of clusters (in number of cells) can either have
  one value (when all clusters have the same height) or as many values as
  there are clusters.} This could for instance be \code{'row = c(2, 52), col
  = 3, width = 4'}, if there are 2 rows of clusters that are 4 wide, starting
  at cell 2,3 and 52,3 \item \code{commodities}: }
}
