% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectData.R
\name{selectData}
\alias{selectData}
\title{Select data chunks from a spreadsheet}
\usage{
selectData(input = NULL, clusters = NULL, header = NULL)
}
\arguments{
\item{input}{the raw table from which to select clusters}

\item{clusters}{the "clusters" slot of a schema description.}

\item{header}{the "header" slot of a schema description.}
}
\description{
Select data chunks from a spreadsheet
}
