% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setHeader.R
\name{setHeader}
\alias{setHeader}
\title{Set the header of a table}
\usage{
setHeader(schema = NULL, rows = NULL, relative = FALSE, merge = FALSE)
}
\arguments{
\item{schema}{[\code{schema(1)}]\cr In case this information is added to an
already existing schema, provide that schema here (overwrites previous
information).}

\item{rows}{[\code{integerish(.)}]\cr The rows in which the header
information are stored.}

\item{relative}{[\code{logical(1)}]\cr Whether or not the values in
\code{rows} are relative to the cluster positions or whether they refer to
the overall table.}

\item{merge}{[\code{logical(1)}]\cr When there is more than one row, this
determines whether or not those rows should be merged.}
}
\value{
An object of class \code{\link{schema}}.
}
\description{
The header of most tables is typically used to provide 'labels' for the data
in that table. With increasingly complex tables, also the header becomes
increasingly complex and this is especially the case in nested tables. This
function describes in which rows the header information are stored and how
they should be treated.
}
\details{
Please also take a look at the currently suggested strategy to set
  up a \link[=schema]{schema description}.
}
\seealso{
Other functions to describe table arrangement: 
\code{\link{setCluster}()},
\code{\link{setFormat}()},
\code{\link{setIDVar}()},
\code{\link{setObsVar}()}
}
\concept{functions to describe table arrangement}
