% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNames.R
\name{setNames}
\alias{setNames}
\title{Set new names to the output table}
\usage{
setNames(temp = NULL, meta = NULL)
}
\arguments{
\item{temp}{a temporary table on which to set names.}

\item{meta}{the schema description that is the basis to derive the names.}
}
\description{
Set new names to the output table
}
