% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{show,schema-method}
\alias{show,schema-method}
\title{Print the \code{schema}}
\usage{
\S4method{show}{schema}(object)
}
\arguments{
\item{object}{[\code{schema}]\cr the schema to print.}
}
\description{
Print the \code{schema}
}
