% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tabs2shift}
\alias{tabs2shift}
\title{List of table types}
\format{
The object of class \code{list} contains 20 different types of tables
  that are used throughout the unit-tests and examples/vignette.
}
\usage{
tabs2shift
}
\description{
List of table types
}
\keyword{datasets}
