% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabshiftr-package.R
\docType{package}
\name{tabshiftr}
\alias{tabshiftr-package}
\alias{tabshiftr}
\title{tabshiftr: Reshape Disorganised Messy Data}
\description{
The tabshiftr package helps the user to build and register schema
descriptions of disorganised (messy) tables. Disorganised tables are tables
that are not in a topologically coherent form, where packages such as
\pkg{tidyr} could be used for reshaping. The schema description documents
the arrangement of input tables and is used to reshape them into a
standardised (tidy) output format.
}
\seealso{
\itemize{
  \item Package website:
        \href{https://luckinet.github.io/tabshiftr/}{https://luckinet.github.io/tabshiftr/}
  \item Github project:
        \href{https://github.com/luckinet/tabshiftr}{https://github.com/luckinet/tabshiftr}
  \item Report bugs:
        \href{https://github.com/luckinet/tabshiftr/issues}{https://github.com/luckinet/tabshiftr/issues}
}
}
\author{
\strong{Maintainer, Author}: Steffen Ehrmann
  \email{steffen.ehrmann@posteo.de}
}
