% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSchema.R
\name{updateSchema}
\alias{updateSchema}
\title{Check schema description for consistency}
\usage{
updateSchema(input = NULL, schema = NULL)
}
\arguments{
\item{input}{an input for which to check a schema description.}

\item{schema}{the schema description.}
}
\description{
Check schema description for consistency
}
