% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateSchema.R
\name{validateSchema}
\alias{validateSchema}
\title{Check schema description for consistency}
\usage{
validateSchema(schema = NULL, input = NULL)
}
\arguments{
\item{schema}{the schema description.}

\item{input}{an input for which to check a schema description.}
}
\value{
An updated schema description that is formally consistent and has all
  .find specs resolved into absolute positions.
}
\description{
This function takes an input table and a respective input schema. It
evaluates all \code{\link{find}} specifications in the scope of the input
table and ensures that the schema is fomally consistent.
}
