% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRowCol.R
\name{.find}
\alias{.find}
\title{Determine row or column on the fly}
\usage{
.find(pattern)
}
\arguments{
\item{pattern}{[\code{character(1)}]\cr character string containing a regular
expression to be matched in the \code{input} rows or columns.}
}
\value{
the index values where \code{pattern} was found.
}
\description{
Find the location of a variable not based on it's columns/rows, but based on
a regular expression/name
}
\details{
This functions is basically a wild-card for when columns or rows
  are not known ad-hoc, but have to be assigned on the fly (for example when
  tables of a series are largely the same, but some variable(s) sit in
  different columns).
}
