% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{.group}
\alias{.group}
\title{Group rows or columns}
\usage{
.group(fn, ...)
}
\arguments{
\item{fn}{[\code{character(1)}]\cr function by which selected columns or rows
shall be combined.}

\item{...}{[\code{integerish(1)}]\cr columns or rows that shall be combined.}
}
\value{
the index values where the target was found.
}
\description{
Group rows or columns
}
