% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_valid.R
\name{expect_valid_table}
\alias{expect_valid_table}
\title{Test for a valid table}
\usage{
expect_valid_table(x = NULL, units = 1)
}
\arguments{
\item{x}{a table to test.}

\item{units}{the number of units in the output table (from 1 to 3)}
}
\value{
Either an error message of the invalid expectations, or the output of
  the last successful expectation.
}
\description{
This function is a collection of expectations which ensure that the output of
\code{\link{reorganise}} is formally and contentwise correct. It is used in
the tests of this package.
}
