% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRowCol.R
\name{find_}
\alias{find_}
\alias{find_row}
\alias{find_col}
\title{Determine row or column from a data-frame}
\usage{
find_row(pattern, input = NULL)

find_col(pattern, input = NULL)
}
\arguments{
\item{pattern}{[\code{character(1)}]\cr character string containing a regular
expression to be matched in the \code{input} rows or columns.}

\item{input}{[\code{data.frame}]\cr the input data-frame to match the
\code{pattern} with.}
}
\value{
the index values where \code{pattern} was found.
}
\description{
Determine row or column from a data-frame
}
\details{
In case \code{input} is not provided or the match fails, the return
  value is \code{NULL}, which is an adaption to these two functions' primary
  use-case as helpers in building a schema description. When piping through
  \code{\link{reorganise}}, the function will be provided with the respective
  input, to determine the specific index values of that data-frame.
}
