% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNames.R
\name{getNames}
\alias{getNames}
\title{get the names from a schema description}
\usage{
getNames(temp = NULL, meta = NULL)
}
\arguments{
\item{temp}{a temporary table for which to derive names.}

\item{meta}{the output of \code{getMetadata} as basis to derive names.}
}
\description{
get the names from a schema description
}
