% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNames.R
\name{getNames}
\alias{getNames}
\title{Determine the names from a schema description}
\usage{
getNames(header = NULL, meta = NULL)
}
\arguments{
\item{header}{the header from which to derive names.}

\item{meta}{the output of \code{getMetadata} as basis to derive names.}
}
\value{
A vector of column names.
}
\description{
This function determines the specific column names that are required in the
process of reshaping.
}
