% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSchema.R
\name{makeSchema}
\alias{makeSchema}
\title{Make a schema description}
\usage{
makeSchema(schema = NULL)
}
\arguments{
\item{schema}{[\code{list(2)}]\cr the list of schema information.}
}
\description{
Make a schema description
}
