% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSchema.R
\name{makeSchema}
\alias{makeSchema}
\title{Make a schema description}
\usage{
makeSchema(schema = NULL)
}
\arguments{
\item{schema}{[\code{list(.)}]\cr the list of schema information. This can
contain the lists \code{clusters}, \code{meta}, \code{header} and
\code{variables}. See \code{\link{schema_default}} for more details.}
}
\description{
Make a schema description
}
