% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSchema.R
\name{makeSchema}
\alias{makeSchema}
\title{Check a schema description}
\usage{
makeSchema(schema = NULL)
}
\arguments{
\item{schema}{[\code{list(.)}]\cr the list of schema information. This can
contain the lists \code{clusters}, \code{format}, \code{header} and
\code{variables}.}
}
\value{
An object of class \code{\link{schema}}.
}
\description{
This function checks whether the schema description is formally correct.
}
\examples{
# define outline of the cluster(s)
theClusters <- list(row = c(1, 8, 8),
                    col = c(1, 1, 4),
                    width = 3,
                    height = 6,
                    id = "territories")

# identify the row(s) where the header is
theHeader <- list(row = 1, rel = TRUE)

# document identifying variables
idVars <- list(
  territories =
    list(type = "id", row = 1, col = 1, rel = TRUE),
  year =
    list(type = "id", row = c(3:6), col = 4, dist = TRUE),
  commodities =
    list(type = "id", col = 1, rel = TRUE))

# document measured variables
measuredVars <- list(
  harvested =
    list(type = "measured", unit = "ha", factor = 1,
         col = 2, rel = TRUE),
  production =
    list(type = "measured", unit = "t", factor = 1,
         col = 3, rel = TRUE))

# make the schema
mySchema <- list(clusters = theClusters,
                 header = theHeader,
                 variables = c(idVars, measuredVars))

makeSchema(schema = mySchema)
}
