% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganise.R
\name{reorganise}
\alias{reorganise}
\title{Rectangrise a table}
\usage{
reorganise(input = NULL)
}
\arguments{
\item{input}{[\code{data.frame(1)}]\cr table to rectangularise.}
}
\description{
This function takes the output from \code{\link{record}} and rearranges
columns and rows so that the resulting table has a perfectly rectangular
format.
}
