% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\docType{class}
\name{schema-class}
\alias{schema-class}
\alias{out}
\title{The \code{schema} class (S4) and its methods}
\description{
A \code{schema} stores the information of where which information is tored in
a table of data.
}
\details{
The slot \code{variables} typically contains several lists that each
record the metadata of the respective variables.
}
\section{Slots}{

\describe{
\item{\code{cluster}}{[\code{named list(6)}]\cr metadata of clusters of data}

\item{\code{variables}}{[\code{named list(.)}]\cr metadata of variables}
}}

