% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\docType{class}
\name{schema-class}
\alias{schema-class}
\alias{schema}
\title{The \code{schema} class (S4) and its methods}
\description{
A \code{schema} stores the information of where which information is stored in
a table of data.
}
\details{
The slot \code{variables} typically contains several lists that each
record the metadata of the respective variables.
}
\section{Slots}{

\describe{
\item{\code{cluster}}{[\code{list(1)}]\cr description of clusters of data.}

\item{\code{header}}{[code{list(1)}]\cr description of the header.}

\item{\code{meta}}{[\code{list(1)}]\cr description of the metadata.}

\item{\code{variables}}{[\code{named list(.)}]\cr description of variables.}
}}

