% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schema_default}
\alias{schema_default}
\title{Default template of a schema description}
\format{
The object of class \code{schema} describes at which position in a
  table which information can be found. It contains the four slots
  \code{clusters}, \code{header}, \code{meta} and \code{variables}.

  There is hardly any limit to how data can be arranged in a spreadsheet,
  apart from the apparent organisation into a lattice of cells. However, it
  is often the case that data are gathered into topologically coherent
  chunks. Those chunks are what is considered 'cluster' in arealDB. Clusters
  are described by the properties: \itemize{ \item \code{row}
  [\code{integerish(1)}]:\cr The vertical cell values of the top-left cell of
  each cluster. This can also be a vector of values in case there are several
  clusters. \item \code{col} [\code{integerish(1)}]:\cr The horizontal cell
  value of the top-left cell of each cluster. This can also be a vector of
  values in case there are several clusters. \item \code{width}
  [\code{integerish(1)}]:\cr The width of each cluster in cells. This can
  also be a vector of values in case there are several clusters \item
  \code{height} [\code{integerish(1)}]:\cr The height of each cluster in
  cells. This can also be a vector of values in case there are several
  clusters \item \code{id} [\code{character(1)}]:\cr When data are clustered,
  it is often the case that the data are segregated according to one of the
  variables of interest. In such cases, this variable needs to be registered
  as cluster ID.}

  The slot \code{header} describes in which rows the header informations are
  stored and how they should be treated. It contains the properties \itemize{
  \item \code{row} [\code{integerish(.)}]\cr The rows in which the header
  information are stored. \item \code{rel} [\code{logical(1)}]\cr  Whether or
  not the values in \code{row} are relative to the cluster positions or
  whether they refer to the overall table. \item \code{merge}
  [\code{logical(1)}]\cr When there is more than one row, this determines
  whether or not those rows should be merged. }

  The slot \code{meta} describes information concerning the values in a
  spreadsheet. It contains the properties \itemize{ \item \code{del}
  [\code{character(.)}]\cr The symbol(s) that are used as delimiter in the
  table to reorganise. \item \code{dec} [\code{character(.)}]\cr The
  symbol(s) that are used as decimal symbol in the table to reorganise. \item
  \code{na} [\code{character(.)}]\cr The symbol(s) that are used as "not
  available" values in the table to reorganise.}

  Each element in the slot \code{variables} is a list that describes one of
  the variables that shall be comprised in the final database. Variables are
  either so-called \emph{identifying variables}, which indicate observation
  units, or \emph{measured variables}, which carry the observed values.
  Identifying variables are described by the properties: \itemize{ \item
  \code{type} [\code{character(1)}]:\cr The value \code{"id"} signals that
  this is an identifying variable. \item \code{col}
  [\code{integerish(1)}]:\cr The column(s) in which the variable values are
  recorded. \item \code{row} [\code{integerish(1)}]:\cr The row(s) in which
  the variable values are recorded. \item \code{split}
  [\code{character(1)}]:\cr A regular expression that matches the part of
  values that are supposed to be part of the variable, when the cells contain
  more than one value (for example seperated by a ","). \item \code{dist}
  [\code{character(1)}]:\cr Whether or not the variable is distinct from a
  cluster. This is the case when the variable is not systematically available
  for all clusters and thus needs to be registered separately from the
  clusters. \item \code{rel} [\code{logical(1)}]:\cr Whether or not the
  values in \code{row} and \code{col} are relative to the cluster positions
  or whether they refer to the overall table. }

  Measured variables are described by the properties: \itemize{ \item
  \code{type} [\code{character(1)}]:\cr the value \code{"values"} signals
  that this is a values variable. \item \code{unit} [\code{character(1)}]:\cr
  The unit in which the values shall be recorded \bold{in the database}.
  \item \code{factor} [\code{character(1)}]:\cr A factor to transform the
  values to \code{unit}. For instance, if values are recorded in acres, but
  shall be contained in the database in hectare, the factor would be 0.40468.
  \item \code{row} [\code{integerish(1)}]:\cr The row(s) in which the
  variable values are recorded. \item \code{col} [\code{integerish(1)}]:\cr
  The column(s) in which the variable values are recorded. \item \code{rel}
  [\code{logical(1)}]:\cr Whether or not the values in \code{row} and
  \code{col} are relative to the cluster positions or whether they refer to
  the overall table. \item \code{key} [\code{character(1)}]:\cr If the
  variable is recorded, together with other variables, so that the variable
  names are listed in one column and the respective values are listed in
  another column, give here the name of the column that contains the variable
  names. \item \code{dist} [\code{character(1)}]:\cr Whether or not the
  variable is distinct from a cluster. This is the case when the variable is
  not systematically available for all clusters and thus needs to be
  registered separately from the clusters. \item \code{value}
  [\code{character(1)}]:\cr If the variable is recorded, together with other
  variables, so that the variable names are listed in one column and the
  respective values are listed in another column, give here the level in the
  names column that stands for the values of this variable. }

  The default schema description contains an example of an identifying and a
  values variables. Further identifying and values variables would be added
  when more variables are contained in a table.
}
\usage{
schema_default
}
\description{
Default template of a schema description
}
\keyword{datasets}
