% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schema_default}
\alias{schema_default}
\title{Default template of a schema description}
\format{
The object of class \code{schema} describes at which position in a
  table which information can be found. It contains the four slots
  \code{clusters}, \code{header}, \code{meta} and \code{variables}.

  The default schema description contains an example of an identifying and a
  values variables. Further identifying and values variables would be added
  when more variables are contained in a table.
}
\usage{
schema_default
}
\description{
Default template of a schema description
}
\keyword{datasets}
