% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schema_default}
\alias{schema_default}
\title{Default template of a schema description}
\format{
The object of class \code{schema} describes at which position in a
  table which information can be found. It contains the four slots
  \code{clusters}, \code{format}, \code{filter} and \code{variables}.

  The default schema description contains all slots and fields that are
  required by default and identifying and observed variables are added to it
  into the \code{variables} slot.
}
\usage{
schema_default
}
\description{
Default template of a schema description
}
\keyword{datasets}
