% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFilter.R
\name{setFilter}
\alias{setFilter}
\title{Set filters to mask a table}
\usage{
setFilter(schema = NULL, columns = NULL, rows = NULL)
}
\arguments{
\item{schema}{[\code{schema(1)}]\cr In case this information is added to an
already existing schema, provide that schema here (overwrites previous
information).}

\item{columns}{[\code{integerish(.)}]\cr the columns to ingore.}

\item{rows}{[\code{integerish(.)}]\cr the rows to ignore.}
}
\description{
This function allows to specify additional rules to (un-)select certain rows
or columns.
}
\details{
Currently the functionality is restricted to filtering out columns
  and rows globally across the whole table.
}
