% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFilter.R
\name{setFilter}
\alias{setFilter}
\title{Set filters}
\usage{
setFilter(schema = NULL, rows = NULL, invert = FALSE)
}
\arguments{
\item{schema}{[\code{schema(1)}]\cr In case this information is added to an
already existing schema, provide that schema here (overwrites previous
information).}

\item{rows}{[\code{integerish(.)}]\cr the rows to ignore.}

\item{invert}{[\code{logical(1)}]\cr whether or not to invert the values
provided in \code{columns} and \code{rows}.}
}
\description{
This function allows to specify additional rules to filter certain rows
}
\details{
Currently the functionality is restricted to masking out columns and
  rows globally across the whole table.
}
