% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFilter.R
\name{setFilter}
\alias{setFilter}
\title{Set filters}
\usage{
setFilter(schema = NULL, rows = NULL, invert = FALSE)
}
\arguments{
\item{schema}{[\code{schema(1)}]\cr In case this information is added to an
already existing schema, provide that schema here (overwrites previous
information).}

\item{rows}{[\code{integerish(.)}]\cr the rows to ignore.}

\item{invert}{[\code{logical(1)}]\cr whether or not to invert the values
provided in \code{columns} and \code{rows}.}
}
\value{
An object of class \code{\link{schema}}.
}
\description{
This function allows to specify additional rules to filter certain rows
}
\examples{
# please check the vignette for examples
}
\seealso{
Other functions to describe table arrangement: 
\code{\link{setCluster}()},
\code{\link{setFormat}()},
\code{\link{setIDVar}()},
\code{\link{setObsVar}()}
}
\concept{functions to describe table arrangement}
