% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSchema.R
\name{updateSchema}
\alias{updateSchema}
\title{Check schema description for consistency}
\usage{
updateSchema(input = NULL, schema = NULL)
}
\arguments{
\item{input}{an input for which to check a schema description.}

\item{schema}{the schema description.}
}
\value{
An updated schema description that fulfills formal requirements to be
  processed by \code{\link{reorganise}}.
}
\description{
This function takes an input table and a respective input schema and ensures
that the cluster specification is complete and that column and row values are
consistent.
}
