% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateInput.R
\name{validateInput}
\alias{validateInput}
\title{Pre-process input table}
\usage{
validateInput(schema = NULL, input = NULL)
}
\arguments{
\item{schema}{[\code{character(1)}]\cr the validated schema description of
\code{input}.}

\item{input}{[\code{character(1)}]\cr table to reorganise.}
}
\value{
a table where columns and rows are grouped and headers are spliced
  into the table.
}
\description{
This function groups rows, splices the header into the table and fills
missing values where they should not exist.
}
